Delete reservation functionality addeded - some fixes

+++ manageDB.php
@@ -24,14 +24,18 @@ function checkDB($link){
 				}
 			}else if($row[0] > $N){
 				//TODO remove machines
-				$query = "SELECT ID FROM machines WHERE ID = (SELECT MAX(ID) FROM machines)";	
+				$query = "SELECT MAX(ID) FROM machines";	
 				for($i = $N; $i < $row[0]; $i++){
 					$res = mysqli_query($link, $query);
+					if(!$res || mysqli_num_rows($res) != 1)
+						die("<h1>Can't find maximun machine ID</h1>");
 					$deleteRow = mysqli_fetch_array($res);
 					$deleteQuery = "DELETE FROM reservations WHERE IDM='".$deleteRow[0]."'";
 					$deleteQuery1 = "DELETE FROM machines WHERE ID='".$deleteRow[0]."'";
-					mysqli_query($link, $deleteQuery);
-					mysqli_query($link, $deleteQuery1);
+					if(!mysqli_query($link, $deleteQuery))
+						die("<h1>Elimination from reservations table failed</h1>");
+					if(!mysqli_query($link, $deleteQuery1))
+						die("<h1>Elimination from machines table failed</h1>");
 					mysqli_free_result($res);
 				}
 				
@@ -43,28 +47,34 @@ function checkDB($link){
 
 function loadDB(){
 	if($link = mysqli_connect('localhost', 'root', '', 'assignment')){
+		//mysqli_report(MYSQLI_REPORT_ERROR);
 		checkDB($link);
-		$query = "SELECT U.Name, U.Surname, M.Name, R.StartTime, R.EndTime 
+		$query = "SELECT U.Name, U.Surname, M.Name, M.ID, R.StartTime, R.EndTime 
 				FROM reservations AS R, users AS U, machines AS M
 				WHERE R.IDU = U.ID AND R.IDM = M.ID
 				ORDER BY R.IDU, R.StartTime";
 		if($res = mysqli_query($link, $query)){
-			echo "<table class=\"reservations_table\">";
-			echo "<th>User</th><th>Machine</th><th>StartTime</th><th>Duration(min)</th>";
-			while ($row = mysqli_fetch_array($res)){
-				$name = $row[0];
-				$surname = $row[1];
-				$machine = $row[2];
-				$duration = $row[4] - $row[3];
-				$startTimeH = intval($row[3]/60);
-				$startTimeHstr = sprintf("%02d", $startTimeH);
-				$startTimeM = $row[3]%60;
-				$startTimeMstr = sprintf("%02d", $startTimeM);
-				echo "<tr>";
-				echo "<td>$name $surname</td><td>$machine</td><td>$startTimeHstr:$startTimeMstr</td><td>$duration</td>";
-				echo "</tr>";
+			if(mysqli_num_rows($res) > 0){
+				echo "<table class=\"reservations_table\">";
+				echo "<th>User</th><th>Machine</th><th>StartTime</th><th>Duration(min)</th>";
+				while ($row = mysqli_fetch_array($res)){
+					$name = $row[0];
+					$surname = $row[1];
+					$machine = $row[2];
+					$machineID = $row[3];
+					$duration = $row[5] - $row[4];
+					$startTimeH = intval($row[4]/60);
+					$startTimeHstr = sprintf("%02d", $startTimeH);
+					$startTimeM = $row[4]%60;
+					$startTimeMstr = sprintf("%02d", $startTimeM);
+					echo "<tr>";
+					echo "<td>$name $surname</td><td>$machine (ID = $machineID)</td><td>$startTimeHstr:$startTimeMstr</td><td>$duration</td>";
+					echo "</tr>";
+				}
+				echo "</table>";
+			} else {
+				echo "<h3 style=\"background-color: #eeeeee; border: 1px solid black; padding: 5px;\">No reservations to be shown</h3>";
 			}
-			echo "</table>";
 			mysqli_free_result($res);
 		} else die("<h1>Can't execute query</h1>");
 		mysqli_close($link);
@@ -73,26 +83,31 @@ function loadDB(){
 
 	function loadUserReservations($username){
 		if($link = mysqli_connect('localhost', 'root', '', 'assignment')){
-			$query = "SELECT M.Name, R.StartTime, R.EndTime
+			$query = "SELECT M.Name, M.ID, R.StartTime, R.EndTime
 				FROM reservations AS R, users AS U, machines AS M
 				WHERE R.IDU = U.ID AND R.IDM = M.ID AND U.Email = '$username'
 				ORDER BY R.StartTime";
 			if($res = mysqli_query($link, $query)){
-				echo "<table class=\"reservations_table\">";
-				echo "<th>Machine</th><th>StartTime</th><th>Duration(min)</th>";
-				while ($row = mysqli_fetch_array($res)){
-					$machine = $row[0];
-					$duration = $row[2] - $row[1];
-					$startTimeH = intval($row[1]/60);
-					$startTimeHstr = sprintf("%02d", $startTimeH);
-					$startTimeM = $row[1]%60;
-					$startTimeMstr = sprintf("%02d", $startTimeM);
-					echo "<tr>";
-					echo "<td>$machine</td><td>$startTimeHstr:$startTimeMstr</td><td>$duration</td>";
-					echo "</tr>";
+					if(mysqli_num_rows($res) > 0){
+					echo "<table class=\"reservations_table\">";
+					echo "<th>Machine</th><th>StartTime</th><th>Duration(min)</th>";
+					while ($row = mysqli_fetch_array($res)){
+						$machine = $row[0];
+						$machineID = $row[1];
+						$duration = $row[3] - $row[2];
+						$startTimeH = intval($row[2]/60);
+						$startTimeHstr = sprintf("%02d", $startTimeH);
+						$startTimeM = $row[2]%60;
+						$startTimeMstr = sprintf("%02d", $startTimeM);
+						echo "<tr>";
+						echo "<td>$machine (ID = $machineID)</td><td>$startTimeHstr:$startTimeMstr</td><td>$duration</td>";
+						echo "</tr>";
+					}
+					echo "</table>";
+				} else {
+					echo "<h3 style=\"background-color: #eeeeee; border: 1px solid black; padding: 5px;\">You have no reservations</h3>";
 				}
-				echo "</table>";
-						mysqli_free_result($res);
+				mysqli_free_result($res);
 			} else die("<h1>Can't execute query</h1>");
 			mysqli_close($link);
 		} else die("<h1>Can't connect to DB</h1>");
